move columnmappings to viewset and serializer
