+++ seed/models/__init__.py
@@ -13,6 +13,7 @@ from .tax_lot_properties import *  # noqa
 from .properties import *  # noqa
 from .tax_lots import *  # noqa
 from .columns import *  # noqa
+from .column_mappings import *  # noqa
 from .column_list_settings import *  # noqa
 from .column_list_settings_columns import *  # noqa
 from .auditlog import *  # noqa
