+++ seed/serializers/column_mappings.py
@@ -13,15 +13,23 @@ from seed.models import ColumnMapping
 from seed.serializers.columns import ColumnSerializer
 
 
-# from seed.serializers.base import ChoiceField
-
-
 class ColumnMappingSerializer(serializers.ModelSerializer):
     organization_id = serializers.PrimaryKeyRelatedField(source='super_organization',
                                                          read_only=True)
-    column_raw = ColumnSerializer()
-    column_mapped = ColumnSerializer()
+    user_id = serializers.PrimaryKeyRelatedField(source='user', read_only=True)
 
     class Meta:
         model = ColumnMapping
-        exclude = ('source_type',)
+        exclude = ('source_type', 'column_raw', 'column_mapped')
+
+    def to_representation(self, obj):
+        """Return only the first items in the column_raw and column_mapped"""
+        result = super(ColumnMappingSerializer, self).to_representation(obj)
+
+        if obj.column_raw and obj.column_raw.first():
+            result['column_raw'] = ColumnSerializer(obj.column_raw.first()).data
+
+        if obj.column_mapped and obj.column_mapped.first():
+            result['column_mapped'] = ColumnSerializer(obj.column_mapped.first()).data
+
+        return result
