+++ seed/serializers/columns.py
@@ -16,6 +16,7 @@ from seed.serializers.base import ChoiceField
 
 
 class ColumnSerializer(serializers.ModelSerializer):
+    name = serializers.SerializerMethodField('concat_name')
     organization_id = serializers.PrimaryKeyRelatedField(source='organization', read_only=True)
     unit_name = serializers.SlugRelatedField(source='unit', slug_field='unit_name', read_only=True)
     unit_type = serializers.SlugRelatedField(source='unit', slug_field='unit_type', read_only=True)
@@ -26,6 +27,13 @@ class ColumnSerializer(serializers.ModelSerializer):
     class Meta:
         model = Column
         fields = (
-            'id', 'organization_id', 'table_name', 'merge_protection', 'shared_field_type',
+            'id', 'name', 'organization_id', 'table_name', 'merge_protection', 'shared_field_type',
             'column_name', 'is_extra_data', 'unit_name', 'unit_type', 'display_name', 'data_type',
         )
+
+    def concat_name(self, obj):
+        """
+        set the name of the column which is a special field because it can take on a
+        relationship with the table_name and have an _extra associated with it
+        """
+        return '%s_%s' % (obj.column_name, obj.id)
