+++ seed/tests/test_api.py
@@ -81,6 +81,7 @@ class SchemaGenerationTests(TestCase):
                  '/api/v2/properties',
                  '/api/v2/labels',
                  '/api/v2/columns',
+                 '/api/v2/column_mappings',
                  '/api/v2/column_list_settings',
                  '/api/v2/green_assessment',
                  '/api/v2/green_assessment_url',
